branch 1.x - apps sync
git-svn-id: svn+ssh://svn.pointclouds.org/pcl/branches/pcl-1.x@6218 a9d63959-f2ad-4865-b262-bf0e56cfafb6

+++ apps/include/pcl/apps/vfh_nn_classifier.h
@@ -247,7 +247,7 @@ namespace pcl
         // compute the VFH feature for this point cloud
         FeatureCloudPtr vfhs = computeFeature (testing_data);
         // compute gaussian parameter producing the desired minimum score (around 50 for the default values)
-        double gaussian_param = -radius / log (min_score);
+        float gaussian_param = - static_cast<float> (radius / log (min_score));
         // TODO accept result to be filled in by reference
         return classifier_.classify(vfhs->points.at (0), radius, gaussian_param);
       }
