+++ apps/src/openni_planar_segmentation.cpp
@@ -61,8 +61,8 @@ class OpenNIPlanarSegmentation
         device_id_ (device_id)
     {
       grid_.setFilterFieldName ("z");
-      grid_.setFilterLimits (0.0, 3.0);
-      grid_.setLeafSize (0.01, 0.01, 0.01);
+      grid_.setFilterLimits (0.0f, 3.0f);
+      grid_.setLeafSize (0.01f, 0.01f, 0.01f);
 
       seg_.setOptimizeCoefficients (true);
       seg_.setModelType (pcl::SACMODEL_PLANE);
@@ -155,7 +155,7 @@ usage (char ** argv)
     for (unsigned deviceIdx = 0; deviceIdx < driver.getNumberDevices (); ++deviceIdx)
     {
       cout << "Device: " << deviceIdx + 1 << ", vendor: " << driver.getVendorName (deviceIdx) << ", product: " << driver.getProductName (deviceIdx)
-              << ", connected: " << (int)driver.getBus (deviceIdx) << " @ " << (int)driver.getAddress (deviceIdx) << ", serial number: \'" << driver.getSerialNumber (deviceIdx) << "\'" << endl;
+              << ", connected: " << driver.getBus (deviceIdx) << " @ " << driver.getAddress (deviceIdx) << ", serial number: \'" << driver.getSerialNumber (deviceIdx) << "\'" << endl;
       cout << "device_id may be #1, #2, ... for the first second etc device in the list or" << endl
            << "                 bus@address for the device connected to a specific usb-bus / address combination (works only in Linux) or" << endl
            << "                 <serial-number> (only in Linux and for devices which provide serial numbers)"  << endl;
